[Feature] Reworking of TextScanner
I've realised that you can actually use the buffered overload TextReader.Read method for much better efficiency. Now I wonder why I didn't do that in the first place.
Also, having a built-in methods for reading fixed blocks and conditional sequences should be more efficient with direct access to the underlying string, as opposed to having an extension method calling single-character Read repeatedly.
It's said that premature optimisation is the root of all evil, but when it comes to parsing potentially large files with potentially complex formats, I'd rather not cut corners on that.